moved scheduler and ceremonies primitives to encointer-primitives
