+++ ceremonies/Cargo.toml
@@ -75,7 +75,7 @@ version = "1.0.101"
 [dependencies.sp-core]
 default-features = false
 package = "sp-core"
-git = "https://github.com/paritytech/substrate.git" 
+git = "https://github.com/paritytech/substrate.git"
 rev = "a208da16"
 
 [dependencies.timestamp]
