+++ primitives/Cargo.toml
@@ -22,3 +22,15 @@ default-features = false
 package = "sp-core"
 git = "https://github.com/paritytech/substrate.git"
 rev = "a208da16"
+
+[dependencies.serde]
+features = ["derive"]
+optional = true
+version = "1.0.101"
+
+[features]
+default = ["std"]
+std = [
+    "codec/std",
+    "serde",
+]
