+++ primitives/src/lib.rs
@@ -1,1 +1,5 @@
+#![cfg_attr(not(feature = "std"), no_std)]
+
+pub mod ceremonies;
 pub mod communities;
+pub mod scheduler;
